made the project app PWA and moved metadata into a sperate file app folder
